+++ src/migrations/m220304_094835_discount_conditions.php
@@ -5,9 +5,9 @@ namespace craft\commerce\migrations;
 use craft\commerce\elements\conditions\addresses\DiscountAddressCondition;
 use craft\commerce\elements\conditions\customers\DiscountCustomerCondition;
 use craft\commerce\elements\conditions\orders\DiscountOrderCondition;
-use craft\commerce\elements\conditions\users\DiscountGroupConditionRule;
 use craft\db\Migration;
 use craft\db\Query;
+use craft\elements\conditions\users\GroupConditionRule;
 use craft\helpers\Json;
 
 /**
@@ -52,24 +52,27 @@ class m220304_094835_discount_conditions extends Migration
                 ->where(['discounts.id' => $id])
                 ->column();
 
+            $userGroupUids = $this->_getGroupUids($discountsUserGroupIds);
+
             $userRules = [];
             if ($discount['userGroupsCondition'] == 'userGroupsAnyOrNone') {
                 // do nothing
             } elseif ($discount['userGroupsCondition'] == 'userGroupsIncludeAll') {
                 foreach ($discountsUserGroupIds as $userGroupId) {
-                    $conditionRule = new DiscountGroupConditionRule();
-                    $conditionRule->setValues($userGroupId);
+                    $conditionRule = new GroupConditionRule();
+                    $userGroup = \Craft::$app->getUserGroups()->getGroupById($userGroupId);
+                    $conditionRule->setValues($userGroup->uid);
                     $conditionRule->operator = 'in';
                     $userRules[] = $conditionRule;
                 }
             } elseif ($discount['userGroupsCondition'] == 'userGroupsIncludeAny') {
-                $conditionRule = new DiscountGroupConditionRule();
-                $conditionRule->setValues($discountsUserGroupIds);
+                $conditionRule = new GroupConditionRule();
+                $conditionRule->setValues($userGroupUids);
                 $conditionRule->operator = 'in';
                 $userRules[] = $conditionRule;
             } elseif ($discount['userGroupsCondition'] == 'userGroupsExcludeAny') {
-                $conditionRule = new DiscountGroupConditionRule();
-                $conditionRule->setValues($discountsUserGroupIds);
+                $conditionRule = new GroupConditionRule();
+                $conditionRule->setValues($userGroupUids);
                 $conditionRule->operator = 'ni';
                 $userRules[] = $conditionRule;
             }
@@ -106,5 +109,18 @@ class m220304_094835_discount_conditions extends Migration
     {
         echo "m220304_094835_discount_conditions cannot be reverted.\n";
         return false;
+    }
+
+    /**
+     * @param array $ids
+     * @return array
+     */
+    private function _getGroupUids(array $ids): array
+    {
+        return array_map(function($value) {
+            $userGroup = \Craft::$app->getUserGroups()->getGroupById(1);
+
+            return $userGroup->uid;
+        }, $ids);
     }
 }